Modify the way multiple inputs/outputs work
Allow for changed detection in individual inputs
when the amount of inputs is equal to the number of outputs.

The out of date inputs/outputs can be retrieved from the 'outdated_inputs'
and 'outdated_outputs' property of the optional Target object passed into
each target function.

This way a target can specify multiple inputs which can be rebuilt
individually when they are out of date with respect to their output,
or if their output file is missing.