Change initialization design, defines
the chdir caused by -C needs to be something that happens
immediately to avoid confusion when doing any logic involving files
outside of targets.

pake.init() does this and registers a leaving directory message on
exit if necessary, it returns a pake.Make object now with the defines
passed in from the command line set on it.  defines are now accessed from
the pake.Make object again.