Make output batching in run_script configurable
The stdout_collect parameter of pake.submake.run_script
can be used to set the preference of having all the scripts
output combined and then written, or written as it is received.

The default value is True, which means all the sub scripts output
is collected together and then written to the file like object
passed in the stdout parameter at once.