Add installable pake script
add pake script that finds pakefile.py or pakefile in the directory
and runs it.

It adds a -f/--file option which can be used more than once to specify
one or more pakefiles to run in order.

setup.py installs this into your environment so you can use it from
the command line by calling "pake"