Replace installable script with entry point
Update documentation to reflect the new way
pakefile scripts are executed.

Installing via pip adds the "pake" command to your
environment, which looks for pakefile.py or pakefile
in the current directory and executes it.

you can also explicitly specify one or more pakefiles
with the -f/--file option (the switch can be used multiple times).

pake will change to the correct directory for each file and run it.