Fix completely broken -f and -C options in pake entry point
-f needs to switch to the directory of the given file and execute, passing any -C option to the pakefile script itself.

When the -C option is different from the files directory, the subscript will handle the next directory change to the -C directory.

===

When -f is not provided, the pake entry point needs to look in the current directory for a pakefile.

If -f is not provided but -C is provided, pake needs to look in the -C directory for a pakefile and switch to that directory if it contains one before executing.

In the above two cases where no -f option is provided, -C/--directory should not be passed to the pakefile script itself.

This is because the pake entry point will always be executing in either the current directory, or the directory specified by -C when no -f options are present.