Handle exit() calls inside task, forward exit code
exit() inside a task now causes pake to exit as soon as it possibly can,
which may not be immediately when multiple jobs are running side by side.

The return code passed to exit() inside of a task becomes the return
code of pake.

===

Write extended documentation for pake.terminate and pake.Pake.terminate
explaining why they should not be used inside of a task.

Provide an example for their actual use case.