Handle pake.TaskException the same way as pake.TaskExitException
Error information and stack trace for pake.TaskException is now printed
at the very end of pake's output instead of inline to the tasks IO queue.

This makes output more consistent when a task encounters an exception, as
well as easier to grok.

===

Updates to documentation pages, elaborate on the correct usage of pake.terminate.

Provide a page explaining how pake reacts to exceptions and calls to exit()
from inside of a task.