Exit with returncodes.PAKEFILE_NOT_FOUND explicitly
parser.error does return the correct return code (2) by coincidence.

But it should be explicit in the code that it's what pake is exiting with
when this condition occurs (missing pakefile)

---

Axe gettext usage for now, no point in using it where it was used unless
most the rest of pake becomes internationalized as well.