Avoid printing general command line responses with pake.Pake.print
The only thing that should be going to the pake.Pake.stdout file
object is output from tasks, and the entering/leaving directory/subpake messages.

Command line responses to bad arguments, or messages such as "Nothing to do" and whatnot should be
printed to pake.conf.stdout and pake.conf.stderr.  Which of these file objects they are printed to
depends on if the message is an error or not.