(Bugfix) don't mutate input lists in _process_i_o_params
This was causing a subtle bug with generators inside of input callables.

Problem found when switching pake.glob to use glob.iglob which returns a generator object.

---

If you used input callables inside of a list for your inputs, and those callables returned
a function which in turn returned a generator object, the callables list would be mutated
when your task ran. (with the previous code in _process_i_o_params).

Which was causing the integration test to fail on the second run in parallel mode,
because one of the tasks for the glob_and_pattern tests was having it's GC pinned
input parameter mutated by the first run.