Remove cyclic dependency check, ignore already visited tasks
Topological sort does not need to throw on already
visited edges, it just needs to ignore them.

Throwing on already visited edges prevents a task from having
dependencies which in turn share one or more of the same dependencies.

Mutual dependencies among sub dependencies is a common thing.

---

The design of pake's task registration API prevents cycles like this:

A -> B -> A

Because B must be registered first to become a dependency of A,
And B cannot depend on A if it is registered before it.

You cannot add dependencies to an already registered task.
There is no means for pre-declaration of a node.
And you will never will be able too since there is no need at all for it.

That makes the cycle above and variations of it impossible.

Topological sorting of the above graph with the current algorithm
would yield this anyway:

B, A

Since it will try to visit B again and ignore it.